# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Hammer CRM API",
        "description": "",
        "contact": {
            "name": "roskus",
            "email": "hello@roskus.com"
        },
        "version": "1.0.2"
    },
    "servers": [
        {
            "url": "/api"
        },
        {
            "url": "http://hammercrm.localhost/api"
        }
    ],
    "paths": {
...
        "/customer/{id}": {
            "get": {
                "tags": [
                    "Customer"
                ],
                "summary": "Get Customer information",
                "operationId": "1cea41a2441b5b00ecfca1bd28619305",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id of the Customer",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Customer found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Customer"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Customer not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/product/{id}": {
            "get": {
                "tags": [
                    "Product"
                ],
                "summary": "Get Product information",
                "operationId": "231948bcc1da3391aec233a0261203d4",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id of Product",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Product found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Product"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Product not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Product"
                ],
                "summary": "Update a Product",
                "operationId": "119f106297a93d02177462c9a32adf68",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id of Product",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Product updated successfully"
                    },
                    "400": {
                        "description": "Bad request, please review the parameters"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Product"
                ],
                "summary": "Delete a Product",
                "operationId": "03480045f76eb685a4448180fe082155",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id of Product",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Product deleted successfully"
                    },
                    "400": {
                        "description": "Bad request, please review the parameters"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/user/{id}": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Get User information",
                "operationId": "1ecd17228eb2a3fb74edcc70ffe2102e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id of User",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "User not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
...
            "Customer": {
                "required": [
                    "name",
                    "country"
                ],
                "properties": {
                    "name": {
                        "description": "Name of the company",
                        "type": "string",
                        "example": "My Company"
                    },
                    "business_name": {
                        "description": "Business name or legal name of the company",
                        "type": "string",
                        "example": "My Company S.A."
                    },
                    "dob": {
                        "description": "Date of fundation of the company",
                        "type": "string",
                        "format": "date",
                        "example": "1990-02-20"
                    },
                    "vat": {
                        "description": "VAT/NIF of the company",
                        "type": "string",
                        "example": "ESX1234567X"
                    },
                    "phone": {
                        "description": "Phone of the company",
                        "type": "string",
                        "example": "+3464500000"
                    },
                    "phone2": {
                        "description": "Phone2 of the company",
                        "type": "string",
                        "example": "+3464500000"
                    }
                },
                "type": "object"
            },
...
            "Product": {
                "required": [
                    "category_id",
                    "name"
                ],
                "properties": {
                    "category_id": {
                        "description": "Category ID of the product",
                        "type": "integer",
                        "example": "1"
                    },
                    "brand_id": {
                        "description": "Brand ID of the product",
                        "type": "integer",
                        "example": "2"
                    },
                    "name": {
                        "description": "Name of the product",
                        "type": "string",
                        "example": "Hervidor Russell Hobbs"
                    },
                    "model": {
                        "description": "Model of the product",
                        "type": "string",
                        "example": "Colours Plus+"
                    },
                    "sku": {
                        "description": "SKU of the product",
                        "type": "string",
                        "example": "RUS*24994-70"
                    },
                    "barcode": {
                        "description": "EAN-13 barcode of the product",
                        "type": "string",
                        "example": "4008496982943"
                    },
                    "photo": {
                        "description": "Photo of the product",
                        "type": "string",
                        "example": "product.jpg"
                    },
                    "cost": {
                        "description": "Cost of the product",
                        "type": "number",
                        "example": "10.5"
                    },
                    "price": {
                        "description": "Price of the product",
                        "type": "number",
                        "example": "21.70"
                    },
                    "min_stock_quantity": {
                        "description": "Min Stock Quantity of the product",
                        "type": "integer",
                        "example": "10"
                    },
                    "quantity": {
                        "description": "Quantity of the product",
                        "type": "integer",
                        "example": "33"
                    },
                    "description": {
                        "description": "Description of the product",
                        "type": "string",
                        "example": "This is a product description"
                    },
                    "elaboration_date": {
                        "description": "Elaboration date of the product",
                        "type": "string",
                        "format": "date",
                        "example": "2023-01-01"
                    },
                    "expiration_date": {
                        "description": "Expiration date of the product",
                        "type": "string",
                        "format": "date",
                        "example": "2023-12-01"
                    },
                    "tags": {
                        "description": "Tags of the product",
                        "type": "string",
                        "example": "['hervidor', 'pava electrica']"
                    }
                },
                "type": "object"
            },
...
            "User": {
                "required": [
                    "first_name",
                    "email",
                    "password"
                ],
                "properties": {
                    "first_name": {
                        "description": "First Name of the user",
                        "type": "string",
                        "example": "John"
                    },
                    "last_name": {
                        "description": "Last Name of the user",
                        "type": "string",
                        "example": "Smith"
                    },
                    "email": {
                        "description": "Email of the user",
                        "type": "string",
                        "format": "email",
                        "example": "john.smith@company.com"
                    },
                    "phone": {
                        "description": "Phone of the user",
                        "type": "string",
                        "example": "+3464500000"
                    },
                    "password": {
                        "description": "Password of the user",
                        "type": "string",
                        "format": "password",
                        "example": ""
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Authorisation with JWT generated tokens",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}